refactor(xml): split formatting helpers
Move XML escaping, JavaScript action rendering, structure rendering, link target formatting, and outline rendering out of the XML formatter so formatXml can focus on assembling document and page sections.
